# src/loaders/swagger/swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Api --> Banco Entre Ríos",
    "description" : "BFF para la web del Banco Entre Rios.  ",
    "contact" : {
      "name" : "banco entre rios",
      "url" : "https://www.bancoentrerios.com.ar/acaparacreditos/#",
      "email" : "roy.cuenca@naturalth.it"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "0.0.3"
  },
  "servers" : [ {
      "url" : "http://localhost:3000/v1/ber",
      "description" : "Local"
    },
    {
      "url" : "https://bancoentrerios.herokuapp.com/v1/ber",
      "description" : "Heroku"
    } 
  ],
  "paths" : {
...
    "/tcma/getCUIL/{dni}" : {
      "get" : {
        "tags" : [ "TCMA" ],
        "description" : "Servicio para Obtener Cuil. Casos de prueba: DNI  = \"11111\" -> El Dni es invalido. DNI = \"11111111\" -> OK. DNI = \"22222222\" -> Servicio no disponible",
        "parameters" : [ {
          "name" : "dni",
          "in" : "path",
          "type":"string",
          "description" : "dni",
          "required" : true,
          "example" : "11111111"
        } ],
        "responses" : {
          "200" : {
            "description" : "Http status code: 200 - OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response OK",
                  "example" : {
                    "error" : false,
                    "message" : "OK",
                    "status" : 200,
                    "payload": {
                      "cuil": "20111111112"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Http status code response: 400 - El DNI es invalido",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response El DNI es invalido",
                  "example" : {
                    "error": {
                        "message": "El DNI es invalido"
                    },
                    "payload": {
                        "isValid": false
                    }
                  }
                }
              }
            }
          },
          "504" : {
            "description" : "Http status code response: 504 - Servicio no disponible",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response Servicio no disponible",
                  "example" :{
                    "error": {
                        "message": "Servicio no disponible"
                    },
                    "payload": {
                        "isValid": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/tcma/getActivities/{cuil}" : {
      "get" : {
        "tags" : [ "TCMA" ],
        "description" : "Servicio para Validar Cuil. Casos de prueba: Cuil  = \"11111\" -> El Cuil es invalido. Cuil = \"11111111111\" -> OK. Cuil = \"22222222222\" -> Servicio no disponible",
        "parameters" : [ {
          "name" : "cuil",
          "in" : "path",
          "type":"string",
          "description" : "cuil",
          "required" : true,
          "example" : "11111111111"
        } ],
        "responses" : {
          "200" : {
            "description" : "Http status code: 200 - OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response OK",
                  "example" : {
                    "error" : false,
                    "message" : "OK",
                    "status" : 200,
                    "payload": [
                      {
                          "value": "000007",
                          "label": "JUBILADO",
                          "code": "212"
                      },
                      {
                        "value": "000008",
                        "label": "ESTUDIANTE",
                        "code": "212"
                      },
                      {
                        "value": "000009",
                        "label": "AMA DE CASA",
                        "code": "321"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Http status code response: 400 - El Cuil es invalido",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response El Cuil es invalido",
                  "example" : {
                    "error": {
                        "message": "El Cuil es invalido"
                    },
                    "payload": {
                        "isValid": false
                    }
                  }
                }
              }
            }
          },
          "504" : {
            "description" : "Http status code response: 504 - Servicio no disponible",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response Servicio no disponible",
                  "example" :{
                    "error": {
                        "message": "Servicio no disponible"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/tcma/getLocalities/{postalCode}" : {
      "get" : {
        "tags" : [ "TCMA" ],
        "description" : "Servicio para Obtener La Localidad. Casos de prueba: Code  = \"111\" -> El Codigo Postal es invalido. Code = \"1426\" -> El Codigo Postal es Valido. Code = \"222\" -> Servicio no disponible",
        "parameters" : [ {
          "name" : "postalCode",
          "in" : "path",
          "type":"string",
          "description" : "postalCode",
          "required" : true,
          "example" : "1426"
        } ],
        "responses" : {
          "200" : {
            "description" : "Http status code: 200 - El Codigo Postal es Valido",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response OK",
                  "example" : {
                    "error" : false,
                    "message" : "OK",
                    "status" : 200,
                    "payload": [ 
                      {
                        "label": "CAPITAL FEDERAL",
                        "postalCodeId": 123,
                        "postalCode": "1426",
                        "provinceId": 1,
                        "province": "CAPITAL FEDERAL",
                        "countryId":58, 
                        "country": "ARGENTINA"
                      }
                    ] 
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Http status code response: 400 - El Codigo Postal es invalido",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response El Codigo Postal es invalido",
                  "example" : {
                    "error": {
                        "message": "El Codigo Postal es invalido"
                    },
                    "payload": {
                        "isValid": false
                    }
                  }
                }
              }
            }
          },
          "504" : {
            "description" : "Http status code response: 504 - Servicio no disponible",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response Servicio no disponible",
                  "example" :{
                    "error": {
                        "message": "Servicio no disponible"
                    },
                    "payload": {
                        "isValid": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tcma/validateApplicantAge/{applicantDayOfBirth}" : {
      "get" : {
        "tags" : [ "TCMA" ],
        "description" : "Servicio para validar una fecha de nacimiento. Casos de prueba: Fecha de nacimiento = \"10%2F13%2F1980\" -> La fecha de nacimiento es invalida. Fecha de nacimiento = \"10%2F01%2F3000\" -> La fecha de nacimiento es mayor a la fecha actual. Fecha de nacimiento = \"10%2F01%2F1980\" -> OK",
        "parameters" : [ {
          "name" : "applicantDayOfBirth",
          "in" : "path",
          "type":"string",
          "description" : "Fecha de nacimiento",
          "required" : true,
          "example" : "10%2F01%2F1980"
        } ],
        "responses" : {
          "200" : {
            "description" : "Http status code: 200 - Fecha valida",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response OK",
                  "example" : {
                    "error" : false,
                    "message" : "OK",
                    "status" : 200,
                    "payload": {
                      "isValid": true
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Http status code response: 400 - Fecha de nacimiento invalida",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response fecha de nacimiento invalida",
                  "example" : {
                    "error": {
                        "message": "La fecha de nacimiento en invalida"
                    },
                    "payload": {
                        "isValid": false
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Http status code response: 400 - Fecha de nacimiento es mayor a la fecha actual",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response Fecha de nacimiento es mayor a la fecha actual",
                  "example" :{
                    "error": {
                        "message": "Fecha de nacimiento es mayor a la fecha actual"
                    },
                    "payload": {
                        "isValid": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tcma/validateCuil/{cuil}" : {
      "get" : {
        "tags" : [ "TCMA" ],
        "description" : "Servicio para Validar Cuil. Casos de prueba: Cuil  = \"11111\" -> El Cuil es invalido. Cuil = \"11111111111\" -> OK. Cuil = \"22222222222\" -> Servicio no disponible",
        "parameters" : [ {
          "name" : "cuil",
          "in" : "path",
          "type":"string",
          "description" : "cuil",
          "required" : true,
          "example" : "11111111111"
        } ],
        "responses" : {
          "200" : {
            "description" : "Http status code: 200 - OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response OK",
                  "example" : {
                    "error" : false,
                    "message" : "OK",
                    "status" : 200,
                    "payload": {
                      "isValid": true
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Http status code response: 400 - El Cuil es invalido",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response El Cuil es invalido",
                  "example" : {
                    "error": {
                        "message": "El Cuil es invalido"
                    },
                    "payload": {
                        "isValid": false
                    }
                  }
                }
              }
            }
          },
          "504" : {
            "description" : "Http status code response: 504 - Servicio no disponible",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Response Servicio no disponible",
                  "example" :{
                    "error": {
                        "message": "Servicio no disponible"
                    },
                    "payload": {
                        "isValid": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tcma/personalData" : {
      "post" : {
        "tags" : [ "TCMA" ],
        "description" : "Endpoint EnvioFormulario Paso 1 (TCMA). Casos de prueba: nombre = Fernando > Http status code response: 400, nombre = Julio > Http status code response: 504, Cualquier otro caso > Http status code response: 400",
        "consumes": "application/json",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/personalData"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Caso ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Http status code response 200",
                  "example" : {
                    "error" : false,
                    "message" : "OK",
                    "status" : 200,
                    "payload" : {
                      "nombre": "luis",
                      "apellido": "fernandez",
                      "dni": "56634553",
                      "email": "luisfernandez@gmail.com",
                      "celularArea": 1154455335,
                      "celularNumero": 1176655443,
                      "telefonoArea": 1154455225,
                      "telefonoNumero": 11533555224,
                      "cuil": 20566345532
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Dato de entrada incorrecto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "nombre = \"Fernando\"",
                  "example" : {
                    "error" : {
                      "message" : "Dato de entrada incorrecto"
                    },
                    "payload" : {
                      "nombre": "fernando",
                      "apellido": "fernandez",
                      "dni": "56634553",
                      "email": "luisfernandez@gmail.com",
                      "celularArea": 1154455335,
                      "celularNumero": 1176655443,
                      "telefonoArea": 1154455225,
                      "telefonoNumero": 11533555224,
                      "cuil": 20566345532
                    } 
                  }
                }
              }
            }
          },
          "504" : {
            "description" : "Servicio del grupo no disponible",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "nombre = \"Julio\"",
                  "example" : {
                    "error" : {
                      "message" : "Servicio del grupo no disponible"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Nombre no encontrado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "nombre = \"cualquiera que no este en los casos\"",
                  "example" : {
                    "error" : {
                      "message" : "Nombre no encontrado"
                    },
                    "payload" : {
                      "nombre": "luis",
                      "apellido": "fernandez",
                      "dni": "56634553",
                      "email": "luisfernandez@gmail.com",
                      "celularArea": 1154455335,
                      "celularNumero": 1176655443,
                      "telefonoArea": 1154455225,
                      "telefonoNumero": 11533555224,
                      "cuil": 20566345532
                    } 
                  }
                }
              }
            }
          }
          
        }
      }
    },
    "/tcma/sentData" : {
      "post" : {
        "tags" : [ "TCMA" ],
        "description" : "Endpoint Envio envio formulario Paso 2 (TCMA). Casos de prueba: sexo = 1 > Http status code response: 200, sexo = 2 > Http status code response: 200, Cualquier otro caso > Http status code response: 400",
        "consumes": "application/json",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/sentData"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Caso ok 1 Masculino",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Http status code response 200",
                  "example" : {
                    "error" : false,
                    "message" : "OK",
                    "status" : 200,
                    "payload" : 
                    {
                      "sexo": 1,
                      "fechaNacimiento": "20%03%1989",
                      "estadoCivil": 1,
                      "idPaisNacionalidad": 1,
                      "idPaisNacimiento": 1,
                      "domicilioCalle": "conde",
                      "domicilioNumero": 1468,
                      "domicilioPiso": 3,
                      "domicilioDpto": "d",
                      "domicilioCP": 1232,
                      "idLocalidad": 1,
                      "codigoActividad": "000009",
                      "personaEP": "2",          
                      "nombre": "luis",
                      "apellido": "fernandez",
                      "dni": "56634553",
                      "email": "luisfernandez@gmail.com",
                      "celularArea": 1154455335,
                      "celularNumero": 1176655443,
                      "telefonoArea": 1154455225,
                      "telefonoNumero": 11533555224,
                      "cuil": 20566345532
                    }
                  }
                }
              }
            }
          },
          "200" : {
            "description" : "Caso ok 2 Femenino",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Http status code response 200",
                  "example" : {
                    "error" : false,
                    "message" : "OK",
                    "status" : 200,
                    "payload" : 
                    {
                      "sexo": 2,
                      "fechaNacimiento": "20%03%1989",
                      "estadoCivil": 1,
                      "idPaisNacionalidad": 1,
                      "idPaisNacimiento": 1,
                      "domicilioCalle": "conde",
                      "domicilioNumero": 1468,
                      "domicilioPiso": 3,
                      "domicilioDpto": "d",
                      "domicilioCP": 1232,
                      "idLocalidad": 1,
                      "codigoActividad": "000009",
                      "personaEP": "2",          
                      "nombre": "luis",
                      "apellido": "fernandez",
                      "dni": "56634553",
                      "email": "luisfernandez@gmail.com",
                      "celularArea": 1154455335,
                      "celularNumero": 1176655443,
                      "telefonoArea": 1154455225,
                      "telefonoNumero": 11533555224,
                      "cuil": 20566345532
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Sexo no especficado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "nombre = \"cualquiera que no este en los casos\"",
                  "example" : {
                    "error" : {
                      "message" : "Sexo no especificado"
                    },
                    "payload" :
                    {
                      "sexo": "no especificado",
                      "fechaNacimiento": "20%03%1989",
                      "estadoCivil": 1,
                      "idPaisNacionalidad": 1,
                      "idPaisNacimiento": 1,
                      "domicilioCalle": "conde",
                      "domicilioNumero": 1468,
                      "domicilioPiso": 3,
                      "domicilioDpto": "d",
                      "domicilioCP": 1232,
                      "idLocalidad": 1,
                      "codigoActividad": "000009",
                      "personaEP": "2",          
                      "nombre": "luis",
                      "apellido": "fernandez",
                      "dni": "56634553",
                      "email": "luisfernandez@gmail.com",
                      "celularArea": 1154455335,
                      "celularNumero": 1176655443,
                      "telefonoArea": 1154455225,
                      "telefonoNumero": 11533555224,
                      "cuil": 20566345532
                    }
                  }
                }
              }
            }
          }
          
        }
      }
    },
...
},
  "components" : {
    "schemas" : {
...
      "personalData" : {
        "type" : "object",
        "properties" : {
          "nombre" : {
            "type" : "string"
          },
          "apellido" : {
            "type" : "string"
          },
          "dni" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "celularArea" : {
            "type" : "number"
          },
          "celularNumero" : {
            "type" : "number"
          },
          "telefonoArea" : {
            "type" : "number"
          },
          "telefonoNumero" : {
            "type" : "number"
          },
          "cuil" : {
            "type" : "number"
          }
        },
        "example" : {
          "nombre": "luis",
          "apellido": "fernandez",
          "dni": "56634553",
          "email": "luisfernandez@gmail.com",
          "celularArea": 1154455335,
          "celularNumero": 1176655443,
          "telefonoArea": 1154455225,
          "telefonoNumero": 11533555224,
          "cuil": 20566345532
        } 
      },
      "sentData" : {
        "type" : "object",
        "properties" : {
          "sexo" : {
            "type" : "number"
          },
          "fechaNacimiento" : {
            "type" : "string"
          },
          "estadoCivil" : {
            "type" : "number"
          },
          "idPaisNacionalidad" : {
            "type" : "number"
          },
          "idPaisNacimiento" : {
            "type" : "number"
          },
          "domicilioCalle" : {
            "type" : "string"
          },
          "domicilioNumero" : {
            "type" : "number"
          },
          "domicilioPiso" : {
            "type" : "number"
          },
          "domicilioDpto" : {
            "type" : "string"
          },
          "domicilioCP" : {
            "type" : "string"
          },
          "idLocalidades" : {
            "type" : "number"
          },
          "codigoActividad" : {
            "type" : "string"
          },
          "personaEP" : {
            "type" : "number"
          },
          "dni" : {
            "type" : "number"
          },
          "nombre" : {
            "type" : "string"
          },
          "apellido" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "celularArea" : {
            "type" : "number"
          },
          "celularNumero" : {
            "type" : "number"
          },
          "telefonoArea" : {
            "type" : "number"
          },
          "telefonoNumero" : {
            "type" : "number"
          },
          "cuil" : {
            "type" : "number"
          }
        },
        "example" : {
          "sexo": 1,
          "fechaNacimiento": "20%03%1989",
          "estadoCivil": 1,
          "idPaisNacionalidad": 1,
          "idPaisNacimiento": 1,
          "domicilioCalle": "conde",
          "domicilioNumero": 1468,
          "domicilioPiso": 3,
          "domicilioDpto": "d",
          "domicilioCP": 1232,
          "idLocalidad": 1,
          "codigoActividad": "000009",
          "personaEP": "2",          
          "nombre": "luis",
          "apellido": "fernandez",
          "dni": "56634553",
          "email": "luisfernandez@gmail.com",
          "celularArea": 1154455335,
          "celularNumero": 1176655443,
          "telefonoArea": 1154455225,
          "telefonoNumero": 11533555224,
          "cuil": 20566345532
        } 
      }, 
...
    }
  }
}